+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## v0.5.0 (00.00.0000)
+## v0.5.0 (08.15.2017)
+ - [NEW] - Added [AnimateCSS](https://daneden.github.io/animate.css/) for every component!
+ - [NEW] - Added a dismissable `Alert` component!
  - [IMPROVED] - Add support for custom units when styling components
 
 ## v0.4.0 (08.13.2017)
- - [NEW] - Chart.js!
+ - [NEW] - [Chart.js](http://www.chartjs.org/)!
  - [FIXED] - Minor issues with `Chart` not rendering correctly